1.3-docker/Task_2/stocks_products/requirements.txt
django
djangorestframework
gunicorn
django-filter
psycopg2-binary